requirements.txt
aiogram==3.15.0
google-genai==0.3.0
aiohttp==3.10.11